Fix #663, Scrub datatypes across TBL
Fix all use of uint32 for sizes, replace with size_t
Scrub all other types to make sure use is consistent,
using proper typedef whenever available.

+++ fsw/cfe-core/src/tbl/cfe_tbl_api.c
@@ -50,7 +50,7 @@
  */
 int32 CFE_TBL_Register( CFE_TBL_Handle_t *TblHandlePtr,
                         const char *Name,
-                        uint32  Size,
+                        size_t  Size,
                         uint16  TblOptionFlags,
                         CFE_TBL_CallbackFuncPtr_t TblValidationFuncPtr )
 {
